add admin req to save/delete/update/getadmin announcements, flyway fixes, pom changes

+++ server-extension/src/main/java/vaylavirasto/announcements/actions/DeleteAnnouncement.java
@@ -17,6 +17,7 @@ public class DeleteAnnouncement extends AnnouncementsRestActionHandler {
     @Override
     public void handlePost(ActionParameters params) throws ActionException {
         requireLiviConfigured();
+        params.requireAdminUser();
 
         try {
             JSONObject result = AnnouncementsDBHelper.deleteAnnouncement(params);
